+++ src/cpptrajdepend
@@ -309,7 +309,7 @@ PDBfile.o : PDBfile.cpp Atom.h CpptrajFile.h CpptrajStdio.h FileIO.h FileName.h 
 PairList.o : PairList.cpp Atom.h AtomExtra.h AtomMask.h AtomType.h Box.h Constants.h CoordinateInfo.h CpptrajStdio.h FileName.h Frame.h MaskToken.h Matrix_3x3.h Molecule.h NameType.h PairList.h Parallel.h ParameterHolders.h ParameterSet.h ParameterTypes.h Range.h ReplicaDimArray.h Residue.h StringRoutines.h SymbolExporting.h Timer.h Topology.h TypeNameHolder.h Vec3.h
 Parallel.o : Parallel.cpp Parallel.h
 ParallelNetcdf.o : ParallelNetcdf.cpp CpptrajStdio.h Parallel.h ParallelNetcdf.h
-ParameterSet.o : ParameterSet.cpp AtomType.h Constants.h CpptrajFile.h CpptrajStdio.h FileIO.h FileName.h NameType.h Parallel.h ParameterHolders.h ParameterSet.h ParameterTypes.h TypeNameHolder.h
+ParameterSet.o : ParameterSet.cpp AtomType.h Constants.h CpptrajFile.h CpptrajStdio.h FileIO.h FileName.h NameType.h Parallel.h ParameterHolders.h ParameterSet.h ParameterTypes.h TypeNameHolder.h UpdateParameters.h
 ParmFile.o : ParmFile.cpp ArgList.h Atom.h AtomExtra.h AtomMask.h AtomType.h BaseIOtype.h BondSearch.h Box.h BufferedFrame.h BufferedLine.h Constants.h CoordinateInfo.h CpptrajFile.h CpptrajStdio.h FileIO.h FileName.h FileTypes.h Frame.h MaskToken.h Matrix_3x3.h Molecule.h NameType.h Parallel.h ParameterHolders.h ParameterSet.h ParameterTypes.h ParmFile.h ParmIO.h Parm_Amber.h Parm_CIF.h Parm_CharmmPsf.h Parm_Gromacs.h Parm_Mol2.h Parm_PDB.h Parm_SDF.h Parm_Tinker.h Range.h ReplicaDimArray.h Residue.h SymbolExporting.h TextFormat.h Topology.h TypeNameHolder.h Vec3.h
 Parm_Amber.o : Parm_Amber.cpp ArgList.h Atom.h AtomExtra.h AtomMask.h AtomType.h BaseIOtype.h BondSearch.h Box.h BufferedFrame.h Constants.h CoordinateInfo.h CpptrajFile.h CpptrajStdio.h FileIO.h FileName.h Frame.h MaskToken.h Matrix_3x3.h Molecule.h NameType.h Parallel.h ParameterHolders.h ParameterSet.h ParameterTypes.h ParmIO.h Parm_Amber.h Range.h ReplicaDimArray.h Residue.h StringRoutines.h SymbolExporting.h TextFormat.h Topology.h TypeNameHolder.h Vec3.h
 Parm_CIF.o : Parm_CIF.cpp ArgList.h Atom.h AtomExtra.h AtomMask.h AtomType.h BaseIOtype.h BondSearch.h Box.h BufferedLine.h CIFfile.h Constants.h CoordinateInfo.h CpptrajFile.h CpptrajStdio.h FileIO.h FileName.h Frame.h MaskToken.h Matrix_3x3.h Molecule.h NameType.h Parallel.h ParameterHolders.h ParameterSet.h ParameterTypes.h ParmIO.h Parm_CIF.h Range.h ReplicaDimArray.h Residue.h StringRoutines.h SymbolExporting.h Topology.h TypeNameHolder.h Vec3.h
@@ -341,7 +341,7 @@ TextFormat.o : TextFormat.cpp StringRoutines.h TextFormat.h
 Timer.o : Timer.cpp CpptrajStdio.h Timer.h
 TinkerFile.o : TinkerFile.cpp ArgList.h Atom.h AtomMask.h Box.h BufferedLine.h CoordinateInfo.h CpptrajFile.h CpptrajStdio.h FileIO.h FileName.h Frame.h MaskToken.h Matrix_3x3.h Molecule.h NameType.h Parallel.h ReplicaDimArray.h Residue.h StringRoutines.h SymbolExporting.h TinkerFile.h Vec3.h
 TopInfo.o : TopInfo.cpp ArgList.h AssociatedData.h Atom.h AtomExtra.h AtomMask.h AtomType.h Box.h CharMask.h Constants.h CoordinateInfo.h CpptrajFile.h CpptrajStdio.h DataSet.h DataSet_Coords.h Dimension.h DistRoutines.h FileIO.h FileName.h Frame.h MaskToken.h Matrix_3x3.h MetaData.h Mol.h Molecule.h NameType.h Parallel.h ParameterHolders.h ParameterSet.h ParameterTypes.h Range.h ReplicaDimArray.h Residue.h StringRoutines.h SymbolExporting.h TextFormat.h TopInfo.h Topology.h TorsionRoutines.h TypeNameHolder.h Vec3.h
-Topology.o : Topology.cpp Atom.h AtomExtra.h AtomMask.h AtomType.h Box.h CharMask.h Constants.h CoordinateInfo.h CpptrajStdio.h FileName.h Frame.h MaskToken.h Matrix_3x3.h Molecule.h NameType.h Parallel.h ParameterHolders.h ParameterSet.h ParameterTypes.h Range.h ReplicaDimArray.h Residue.h StringRoutines.h SymbolExporting.h Topology.h TypeNameHolder.h Vec3.h
+Topology.o : Topology.cpp Atom.h AtomExtra.h AtomMask.h AtomType.h Box.h CharMask.h Constants.h CoordinateInfo.h CpptrajStdio.h FileName.h Frame.h MaskToken.h Matrix_3x3.h Molecule.h NameType.h Parallel.h ParameterHolders.h ParameterSet.h ParameterTypes.h Range.h ReplicaDimArray.h Residue.h StringRoutines.h SymbolExporting.h Topology.h TypeNameHolder.h UpdateParameters.h Vec3.h
 TorsionRoutines.o : TorsionRoutines.cpp Constants.h TorsionRoutines.h Vec3.h
 TrajFrameCounter.o : TrajFrameCounter.cpp ArgList.h CpptrajStdio.h TrajFrameCounter.h
 TrajIOarray.o : TrajIOarray.cpp ArgList.h Atom.h AtomExtra.h AtomMask.h AtomType.h BaseIOtype.h Box.h Constants.h CoordinateInfo.h CpptrajStdio.h FileName.h FileTypes.h Frame.h FramePtrArray.h MaskToken.h Matrix_3x3.h Molecule.h NameType.h Parallel.h ParameterHolders.h ParameterSet.h ParameterTypes.h Range.h ReplicaDimArray.h Residue.h StringRoutines.h SymbolExporting.h Topology.h TrajFrameCounter.h TrajIOarray.h TrajectoryFile.h TrajectoryIO.h TypeNameHolder.h Vec3.h
